(#237)refactor: 방문자 service interface 구현했습니다.
